+++ CHANGES.rst
@@ -1,6 +1,9 @@
 1.2 (unreleased)
 ================
 
+- ``TheJoker`` now expects a ``numpy.random.Generator`` instance instead of a
+  ``RandomState`` for controlling random number generation.
+
 1.1 (2020-04-19)
 ================